pyproject.toml
[project]
name = "openqa-log-visualizer"
version = "0.1.0"
description = "A web-based tool to analyze and visualize logs from multiple, related openQA jobs."
requires-python = ">=3.9"
license = "GPL-2.0-only"
readme = "README.md"

dependencies = [
    "flask>=3.1.2",
    "openqa-client>=4.3.0",
    "pyyaml",
    "requests",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
